Migrate pure-logic tests from tests/ to unit tests (Phase 2)
Moves conversion, parsing, icon-selection, formatting, and error-priority
tests that exercised no I/O or public-API boundary out of tests/ (where
each file compiles as a separate crate) into #[cfg(test)] modules
co-located with the code they test, per docs/test-suite-review.md.

- BOM/Open-Meteo response deserialization -> apis::bom::models,
  apis::open_meteo::models (also tightens plausibility-range assertions
  to exact fixture values)
- from_bom/from_open_meteo conversion, DST handling, wind speed, snow
  detection -> domain::models
- icon_name boundary logic -> domain::icons
- date format validation -> configs::validation
- ContextBuilder error-priority display, daily forecast date mapping
  (NY timezone, Issue #16 regression) -> dashboard::context

Deletes tests/clock_integration_test.rs outright: it tested chrono
itself and a function defined inline in the test, not application code.

Full suite: 185 passed before and after; test count is preserved
exactly (each deleted integration test file's assertions reappear as
unit tests). tests/ now holds only genuine integration tests
(wiremock-backed HTTP/retry tests, full-pipeline snapshot tests).
